Added warning messages for update and delete

+++ frontend/src/Todo.jsx
@@ -1,36 +1,69 @@
+import Swal from 'sweetalert2';
+
 export default function Todo(props) {
     const { todo, setTodos } = props;
 
     const updateTodo = async (todoId, todoStatus) => {
-        const res = await fetch(`/api/todos/${todoId}`, {
-            method: "PUT",
-            body: JSON.stringify({ status: todoStatus }),
-            headers: {
-                "Content-Type": "application/json"
-            }
-        });
-        const json = await res.json();
-        if (json.acknowledged) {
-            setTodos(currentTodos => {
-                return currentTodos.map((currentTodo) => {
-                    if (currentTodo._id === todoId) {
-                        return { ...currentTodo, status: !currentTodo.status }; //it creates a new object that: Copies all properties (...currentTodo)., Overrides the status property with the new value (!currentTodo.status)
-                    }
-                    return currentTodo;
-                });
+        let userConfirmed = false;
+
+        if (todoStatus === false) {
+          userConfirmed = await Swal.fire({
+            title: 'Are you sure you have completed it?',
+            icon: 'question',
+            showCancelButton: true,
+            confirmButtonText: 'Yes',
+            cancelButtonText: 'No'
+          }).then((result) => result.isConfirmed);
+        } else {
+          userConfirmed = await Swal.fire({
+            title: 'Are you sure you want to revoke the completion status?',
+            icon: 'question',
+            showCancelButton: true,
+            confirmButtonText: 'Yes',
+            cancelButtonText: 'No'
+          }).then((result) => result.isConfirmed);
+        }
+        
+        if (userConfirmed) {
+            const res = await fetch(`/api/todos/${todoId}`, {
+                method: "PUT",
+                body: JSON.stringify({ status: todoStatus }),
+                headers: {
+                    "Content-Type": "application/json"
+                }
             });
+            const json = await res.json();
+            if (json.acknowledged) {
+                setTodos(currentTodos => {
+                    return currentTodos.map((currentTodo) => {
+                        if (currentTodo._id === todoId) {
+                            return { ...currentTodo, status: !currentTodo.status }; //it creates a new object that: Copies all properties (...currentTodo)., Overrides the status property with the new value (!currentTodo.status)
+                        }
+                        return currentTodo;
+                    });
+                });
+            }
         }
     };
 
     const deleteTodo = async (todoId) => {
-        const res = await fetch(`/api/todos/${todoId}`, {
-            method: "DELETE",
-        });
-        const json = await res.json();
-        if (json.acknowledged) {
-            setTodos(currentTodos => {
-                return currentTodos.filter((currentTodo) => (currentTodo._id !== todoId));
-            })
+        const userConfirmed = await Swal.fire({
+            title: 'Are you sure you want to delete this task?',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Yes',
+            cancelButtonText: 'No'
+          }).then((result) => result.isConfirmed);
+        if (userConfirmed){
+            const res = await fetch(`/api/todos/${todoId}`, {
+                method: "DELETE",
+            });
+            const json = await res.json();
+            if (json.acknowledged) {
+                setTodos(currentTodos => {
+                    return currentTodos.filter((currentTodo) => (currentTodo._id !== todoId));
+                })
+            }
         }
     };
